Use node IP lists for HA Proxy ipaddresses
This patch updates the loadbalancer class so that it
defaults to trying to use the node IP list for each
respective service. This data is provided via Hiera
directly (all-nodes-config provides it via the Heat
templates).

By default the ctlplane IP address list is still used
if no service node IP list is provided.

Change-Id: I34cbdf8bd525e6ab61859fe8b8c18fe613dabbfe