Enable X-Forwarded-Proto header for keystone_public
One of the ways to make use of TLS in keystone is through the usage of
the X-Fowarded-Proto header, which will be forwarded with the request
by the loadbalancer, and it will tell keystone what protocol was used
to access it. This also requires configuration from the keystone side.

Change-Id: I9b899ba95e28b7dfae0c1ed84ca8431054673925
(cherry picked from commit 846b4fe0b82f861c3f6c637cdc5d7d9f052e4338)