Override X-Forwarded-Proto header
Right now, the only manipulation done to the X-Forwarded-Proto header
is done if an SSL connection is established. This is not sufficient as
one might be able to erroneously put values through that header.

This patch disables that behaviour by defaulting to plain http if an
SSL connection is not established.

Change-Id: I4bf6def21e21148834c2baa9669190bab8fa95ef
(cherry picked from commit 5f915a88e7c4eec6e6a0341808f0fbc4a4b1a161)