IPv6 dual-stack support
TL;DR:
If keystone_public_api_vip and/or public_virtual_ip is an array of IPs,
HAproxy will be configured to listen on all IPs that are given in the
arrays.
It allows to specify an array for keystone_public_api_vip and/or
public_virtual_ip where one IP is v4 and another one is v6.
HAproxy will configured to listen on both and redirect the traffic to
the IPv6 network (Dual-Stack).

Implementation & background:
HAproxy requires binding options as an hash where each IP contains an
array of binding options.
TripleO does not support Puppet Parser [1] (yet) so we can't manipulate
data iterations inside the manifests.
This patch creates a custom function, called list_to_hash.

Example:
keystone_vips = ['192.168.0.1:5000', '192.168.0.2:5000']
$keystone_bind_opts = ['transparent']

Using this function:
$keystone_vips_hash = list_to_hash($keystone_vips,
$keystone_bind_opts)

Would return:
$keystone_vips_hash = {
  '192.168.0.1:5000' => ['transparent'],
  '192.168.0.2:5000' => ['transparent'],
}

This function will help us in loadbalancer.pp to construct binding
options in dynamic way.
It's backward compatible, so you don't have to give an array.
But if you do, multiple binding will be configured in HAproxy and you'll
also be able to deploy IPv6 Dual-Stack.

[1] https://docs.puppetlabs.com/puppet/latest/reference/lang_iteration.html

Change-Id: I003b6d7d171652654745861d4231882f9e0d373e