Fix comparison to control_virtual_ip
When managing the vip's, we were incorrectly comparing the vip to
$control_virtual_interface instead of $controller_virtual_ip when
determining if we needed to actually create the vip or not.

This caused the vips for internal api, storage, and storage mgmt to
always be created even if they were the same as the control vip. Afaict,
this didn't actually cause any problems, other than having extra vip's
created when they weren't needed. Still, this corrects the code to do
what it was intended to do.

Change-Id: I29aee95afcba25008b8b7bee37ba636eb2595cca
(cherry picked from commit 4c0a136a40655171a14ca4be304193863eac06bd)
Closes-Bug: #1602319