Add missing services ports to service_ports map
AODH, Gnocchi, Sahara and Trove were missing from the service_ports
maps and thus had hardcoded ports in the listener configuration. The
addition of those ports to the map is required to give the
possibility to deployers to configure those ports if needed. This
commit adds them to that map.

Change-Id: Id009d65bf68ba91f97b0d60d32028da50fc88fc3