keystone: fix a race condition in bootstrap
Before, we had bootstrap_master set to 'undef' by default that was
used to whether or not run keystone bootstrap exec during deployment.

Setting the value to undef was a mistake, because enable_bootstrap in
puppet-keystone is set to true by default, so bootstrap was running
on all controllers at step 4 for HA scenario, while we want it to run on
a single controller (pacemaker_master) at step 4, like we do for
db-sync.

This patch:

* removes bootstrap_master.
* re-use sync_db to whether or not run keystone bootstrap.
  so it will only run on a single node when database is ready and
  db-sync done (orchestration dones by anchors in puppet-keystone).

Change-Id: I1042862f7c346d1c358b908c33eae0f33afd5e9f