Add ironic inspector as a terminated HAProxy endpoint
This is needed for the undercloud, as it's in HAProxy where we make
the SSL terminations.

It's not possible to cherry-pick due to the changes we've done in the
master branch. This is because in stable/mitaka we don't have this
haproxy.pp manifest, instead we had loadbalancer.pp. And on the other
hand, we now have a puppet resource that make it easy to add new
endpoints. This resource (::tripleo::haproxy::endpoint) doesn't exist
in stable/mitaka.

Change-Id: Ie4d652b4e5a95849c2fa32a5ce5ecec09ccb6bd9
Related-Bug: #1595047