Nova API and Neutron server should run database creation in step 3
We perform the Galera setup in step 2 so there is no guarantee that the
database will be available in that same step [1].

We used to implement a dependency in puppet using the 'galera-ready'
resource (clustercheck) but this is not possible with roles because we
also don't have any guarantee about clustercheck being installed on the
same node.

Because of the above all services must create/sync their databases
in a later step. This patch fixes Nova API and Neutron Server, the other
services use step 3 already.

1. https://github.com/openstack/tripleo-heat-templates/blob/master/puppet/services/README.rst

Change-Id: I22750ffb64afbe40b5560a6a0d0dabc5b8927d32