use parameter to lookup the step instead of hiera again
In some profiles, we were looking up the $step by using Hiera
again, while we already do it in the parameter definition.
When using this class outside THT, it will fail but with this patch, we
could use just set the $step parameter and the rest of the manifest will
work.

Change-Id: I7082f47204fb4e529b164e4c4f1032e7bdd88f02