Run local CA trust before haproxy deployment
Before haproxy tries to use the TLS certificates it should already
trust the CA. So it's necessary for the local CA-related manifest to
notify the ::tripleo::haproxy class.

This works for newly set deployments. deployments that have already
ran the ca-trust section will already trust the CA and thus won't
need that part.

Change-Id: I32ded4e33abffd51f220fb8a7dc6263aace72acd