Fix parameters and headers inconsistency in the puppet manifests.
As we are staring to manually check overcloud services
the first step is to check that the puppet profiles
are all aligned.

Changes applied:

  No logic added or removed in this submission.

  Removed unused parameters.

  Align header comments structure.

  All profiles parameters sorted following:
  "Mandatory params first sorted alphabetically
  then optional params sorted alphabetically."

Note: Following submissions will check pacemaker,
cinder, mistral and redis services in the base profiles
as some of them has the $pacemaker_master parameter
defaulted to true.

Change-Id: I2f91c3f6baa33f74b5625789eec83233179a9655