Added X-Forwarded-Proto headers for horizon
When enabling federated authentication with keystone, and then enabling websso
in horizon, the URL horizon constructs for the redirect is done internally, and
django needs to be able to know if it has to construct the url with http or
https. By setting this header at the haproxy level, horizon can make the correct
decision.

Change-Id: I0281fe1e5efa0d3f5983342dec70752246d9fca8