Create memcache_servers list in swift proxy profile
Instead of mangling this list in t-h-t, generate the list derived
from memcached_node_ips, which is now always set when memcached
is deployed, regardless of the role.

Note the port default is hard-coded as this is already hard-coded
(in two places) in t-h-t, but we can override it if this changes
in future.

We need this to remove the swift specific stuff out of overcloud.yaml
to enable custom-roles.

Change-Id: Ic8872e5e51732874ca5b93bff5efd3e7ed75bc31
Partially-Implements: blueprint custom-roles