gnocchi: move statsd and metricd at step5
Gnocchi statsd and metricd require Keystone resources to be in place
when using Swift backend, because those services will try to access to
Swift containers.

To do so, we want to move the service start at step 5 instead of 4 and
also require Keystone resource to be managed before starting the
services.

Change-Id: Ie5bc1481a8700c7cd080a76d0978146a84825767
Closes-Bug: #1621164