Move inclusion of ::manila::db::mysql in manila/api profile
In puppet-manila it is the api service performing db sync, not
scheduler. This change moves ::manila::db::mysql (which creates
the empty database and users) in the tripleo manila/api profile.

Also moves rabbit config into a general manila base profile as
that would be needed by the scheduler service as well.

Change-Id: I2b537f735b8d1be8f39e8c274be3872b193c1014