packages: run upgrade at 'setup' stage
Instead of using an operator to make sure we upgrade package before any
service, which causes dependency cycles with iptables puppet module,
let's do another approach where we upgrade rpms in the 'setup' stage,
which is a stage that runs before configure and running services.

In that way, we'll remove dependency cycles and make sure packages are
upgrades before configure and running TripleO services.

Change-Id: I1be83f88be1959885c980ab4f428477d412751f7