Remove faulty migration logic to stop nova-api
The patch making nova run over httpd had added migration logic to
stop nova-api, However, this doesn't work since nova-metadata is
running over the same process. Now, the fact that is was running
seems to be just luck, since the systemctl runs, then we start the
service via the nova::api resource. So this is fragile in it's
current state.

This then removes the exec, as we don't need it for the migration.

Change-Id: I4603b81d30a704b07eef461b3cdbfe164614b04f